third_party/blink/web_tests/http/tests/eventsource/eventsource-eof-expected.txt

Test that EventSource discards event data if there is no newline before eof.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS evt.type is "message"
PASS evt.data is "DATA1"
PASS evt.lastEventId is "1"
PASS evt.type is "message"
PASS evt.data is "DATA2"
PASS evt.lastEventId is "2"
PASS evt.type is "message"
PASS evt.data is "DATA3"
PASS evt.lastEventId is "3"
PASS evt.type is "msg"
PASS evt.data is "DATA"
PASS evt.lastEventId is "3.1"
PASS successfullyParsed is true

TEST COMPLETE
